Units can now be ordered to form hedgehogs
The function is called on a unit, with a center and a radius. Each
soldier is then given a coordinate to proceed towards, which he does at
his speed.

Also, there's a pretty background of France. Google Maps of some random
location approx along De Gaulle's counterattack of 17-19 May.

to-do: fix icons (arrows aren't working for me for some reason)
to-do: add line-based formation maneuvers
to-do: add opposing army and copy interaction code