Implimented squads
The entire army on each side is now composed of a single unit. This unit
can have variable numbers of each type of troop. Each type of troop has
its own stats and attacks individually. Losses are applied on the
attackers turn after the attacker has fired. This means that the person
that goes first doesn't have as much of an advantage. Still working on
calculating deaths. The stats will need updated but the basis is all
there.