Que fully functioning
Units now que up perfectly. They will fill in the nearest spot to the
bridge whenever it becomes available for any reason. This includes the
best spots in the back and front line, not just the bridge adjacent
spot. They also no longer use a que of booleans to figure out if a spot
is free. They instead communicate with each other to figure out where
the others want to go and don't take a spot someone else is looking to
take.