Bridge crossing complete
All of the elements of crossing the bridge have been completed. The que
has been improved. Units will only go to spots closer than their current
one. The time it takes to get across the bridge is accurate to our
mathematical model. The number of soldiers in a units determines how
long it takes to get across. The only thing not implimented is spreading
shots between all enemies in range.