Any soldiers to any bridge
You can now send any or all units to all of the bridges. You can have
none to a bridge or all. It doesn't matter. The groups have been moved
closer to the river. Currently the connecting units between the groups
are being worked on. A great deal of excess code has been reworked and
improved.